Add interfaces in service logic; add logic to delete artists; modify searching for event picture by id instead of event name; remove unused methods in controllers; removed path keywords in mapping annotations
