+++ src/main/java/com/example/eventinformationsystembackend/controller/ArtistController.java
@@ -2,10 +2,8 @@ package com.example.eventinformationsystembackend.controller;
 
 import com.example.eventinformationsystembackend.dto.ArtistDto;
 import com.example.eventinformationsystembackend.dto.ArtistDtoResponse;
-import com.example.eventinformationsystembackend.dto.UserDtoResponse;
-import com.example.eventinformationsystembackend.dto.UserUpdateDto;
-import com.example.eventinformationsystembackend.model.Artist;
-import com.example.eventinformationsystembackend.service.ArtistService;
+import com.example.eventinformationsystembackend.exception.RequiredPictureMissingException;
+import com.example.eventinformationsystembackend.service.implementation.ArtistServiceImpl;
 import jakarta.validation.Valid;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
@@ -17,25 +15,27 @@ import org.springframework.web.multipart.MultipartFile;
 import java.io.IOException;
 import java.util.List;
 
+import static com.example.eventinformationsystembackend.common.ExceptionMessages.REQUIRED_PICTURE_IS_MISSING;
+
 @RestController
-@RequestMapping(path = "/artists")
+@RequestMapping("/artists")
 public class ArtistController {
-    private final ArtistService artistService;
+    private final ArtistServiceImpl artistServiceImpl;
 
     @Autowired
-    public ArtistController(ArtistService artistService) {
-        this.artistService = artistService;
+    public ArtistController(ArtistServiceImpl artistServiceImpl) {
+        this.artistServiceImpl = artistServiceImpl;
     }
 
     @GetMapping()
     public List<ArtistDtoResponse> getAllArtists() {
-        return artistService.getAllArtists();
+        return artistServiceImpl.getAllArtists();
     }
 
-    @GetMapping(path = "/profile-picture/{id}")
+    @GetMapping("/profile-picture/{id}")
     public ResponseEntity<?> getArtistProfilePicture(
-            @PathVariable("id") Long id) throws IOException {
-        byte[] profilePicture = artistService.getArtistProfilePicture(id);
+            @PathVariable Long id) throws IOException {
+        byte[] profilePicture = artistServiceImpl.getArtistProfilePicture(id);
         return ResponseEntity.status(HttpStatus.OK)
                 .contentType(MediaType.valueOf("image/png"))
                 .body(profilePicture);
@@ -44,13 +44,23 @@ public class ArtistController {
     @PostMapping
     public ArtistDtoResponse addArtist(@RequestPart @Valid ArtistDto artistDto,
                                        @RequestPart(value = "profilePicture") MultipartFile profilePicture) {
-        return artistService.addArtist(artistDto, profilePicture);
+        if (profilePicture == null) {
+            throw new RequiredPictureMissingException(REQUIRED_PICTURE_IS_MISSING);
+        }
+
+        return artistServiceImpl.addArtist(artistDto, profilePicture);
     }
 
-    @PutMapping(path = "/{id}")
-    public ArtistDtoResponse updateArtist(@PathVariable("id") Long id,
+    @PutMapping("/{id}")
+    public ArtistDtoResponse updateArtist(@PathVariable Long id,
                                           @RequestPart @Valid ArtistDto artistDto,
                                           @RequestPart(required = false) MultipartFile profilePicture) {
-        return artistService.updateArtist(id, artistDto, profilePicture);
+        return artistServiceImpl.updateArtist(id, artistDto, profilePicture);
+    }
+
+    @DeleteMapping("/{id}")
+    public ResponseEntity<Void> deleteArtist(@PathVariable Long id) {
+        artistServiceImpl.deleteArtist(id);
+        return ResponseEntity.noContent().build();
     }
 }
