+++ src/main/java/com/example/eventinformationsystembackend/controller/CartItemController.java
@@ -2,63 +2,60 @@ package com.example.eventinformationsystembackend.controller;
 
 import com.example.eventinformationsystembackend.dto.CartItemDto;
 import com.example.eventinformationsystembackend.dto.CartItemDtoResponse;
-import com.example.eventinformationsystembackend.model.CartItem;
-import com.example.eventinformationsystembackend.service.CartItemService;
-import com.example.eventinformationsystembackend.service.OrderService;
+import com.example.eventinformationsystembackend.service.implementation.CartItemServiceImpl;
+import com.example.eventinformationsystembackend.service.implementation.OrderServiceImpl;
 import jakarta.validation.Valid;
-import org.apache.coyote.Response;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
-import java.util.Set;
 
 @RestController
-@RequestMapping(path = "/cart")
+@RequestMapping("/cart")
 public class CartItemController {
 
-    private final CartItemService cartItemService;
-    private final OrderService orderService;
+    private final CartItemServiceImpl cartItemServiceImpl;
+    private final OrderServiceImpl orderServiceImpl;
 
     @Autowired
-    public CartItemController(CartItemService cartItemService, OrderService orderService) {
-        this.cartItemService = cartItemService;
-        this.orderService = orderService;
+    public CartItemController(CartItemServiceImpl cartItemServiceImpl, OrderServiceImpl orderServiceImpl) {
+        this.cartItemServiceImpl = cartItemServiceImpl;
+        this.orderServiceImpl = orderServiceImpl;
     }
 
-    @GetMapping(path = "/coupon/{username}")
-    public Double applyCoupon(@PathVariable("username") String username,
+    @GetMapping("/coupon/{username}")
+    public Double applyCoupon(@PathVariable String username,
                               @RequestParam(value = "couponCode") String couponCode) {
-        return orderService.getCouponDiscount(username, couponCode);
+        return orderServiceImpl.getCouponDiscount(username, couponCode);
     }
 
     @GetMapping("/number/{username}")
     public Integer getAllCartItemsNumberForUser(
-            @PathVariable("username") String username) {
-        return cartItemService.getAllCartItemsNumberForUser(username);
+            @PathVariable String username) {
+        return cartItemServiceImpl.getAllCartItemsNumberForUser(username);
     }
 
     @GetMapping("/{username}")
     public List<CartItemDtoResponse> getAllCartItemsForUser(
-            @PathVariable("username") String username) {
-        return cartItemService.getAllCartItemsForUser(username);
+            @PathVariable String username) {
+        return cartItemServiceImpl.getAllCartItemsForUser(username);
     }
 
     @PostMapping
     public CartItemDtoResponse addCartItem(@RequestBody @Valid CartItemDto cartItemDto) {
-        return cartItemService.addCartItem(cartItemDto);
+        return cartItemServiceImpl.addCartItem(cartItemDto);
     }
 
-    @PostMapping(path = "/decrease")
+    @PostMapping("/decrease")
     public ResponseEntity<Void> decreaseCartItemTicketQuantity(@RequestBody @Valid CartItemDto cartItemDto) {
-        cartItemService.decreaseCartItemTicketQuantity(cartItemDto);
+        cartItemServiceImpl.decreaseCartItemTicketQuantity(cartItemDto);
         return ResponseEntity.noContent().build();
     }
 
-    @DeleteMapping(path = "/{id}")
+    @DeleteMapping("/{id}")
     public ResponseEntity<Void> removeCartItem(@PathVariable("id") Long id) {
-        cartItemService.removeCartItem(id);
+        cartItemServiceImpl.removeCartItem(id);
         return ResponseEntity.noContent().build();
     }
 }
