+++ src/main/java/com/example/eventinformationsystembackend/controller/CommentController.java
@@ -2,55 +2,52 @@ package com.example.eventinformationsystembackend.controller;
 
 import com.example.eventinformationsystembackend.dto.CommentDto;
 import com.example.eventinformationsystembackend.dto.CommentDtoResponse;
-import com.example.eventinformationsystembackend.service.CommentService;
+import com.example.eventinformationsystembackend.service.implementation.CommentServiceImpl;
 import jakarta.validation.Valid;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.http.HttpEntity;
 import org.springframework.http.ResponseEntity;
-import org.springframework.security.config.annotation.web.builders.HttpSecurity;
 import org.springframework.web.bind.annotation.*;
 
-import java.net.http.HttpResponse;
 import java.util.List;
 
 @RestController
-@RequestMapping(path = "/comments")
+@RequestMapping("/comments")
 public class CommentController {
-    private final CommentService commentService;
+    private final CommentServiceImpl commentServiceImpl;
 
     @Autowired
-    public CommentController(CommentService commentService) {
-        this.commentService = commentService;
+    public CommentController(CommentServiceImpl commentServiceImpl) {
+        this.commentServiceImpl = commentServiceImpl;
     }
 
-    @GetMapping(path = "/{username}")
+    @GetMapping("/{username}")
     public List<CommentDtoResponse> getAllCommentsByUser(
-            @PathVariable("username") String username) {
-        return commentService.getAllCommentsByUser(username);
+            @PathVariable String username) {
+        return commentServiceImpl.getAllCommentsByUser(username);
     }
 
-    @PatchMapping(path = "/{commentId}/is-read")
-    public ResponseEntity<Void> markCommentAsRead(@PathVariable("commentId") Long commentId) {
-        commentService.markCommentAsRead(commentId);
+    @PatchMapping("/{id}/is-read")
+    public ResponseEntity<Void> markCommentAsRead(@PathVariable Long id) {
+        commentServiceImpl.markCommentAsRead(id);
         return ResponseEntity.ok().build();
     }
 
-    @PatchMapping(path = "/{commentId}/is-removed")
-    public ResponseEntity<Void> markCommentAsRemoved(@PathVariable("commentId") Long commentId) {
-        commentService.markCommentAsRemoved(commentId);
+    @PatchMapping("/{id}/is-removed")
+    public ResponseEntity<Void> markCommentAsRemoved(@PathVariable Long id) {
+        commentServiceImpl.markCommentAsRemoved(id);
         return ResponseEntity.ok().build();
     }
 
-    @PostMapping(path = "/{username}/{postId}")
-    public CommentDtoResponse addComment(@PathVariable("username") String username,
-                                         @PathVariable("postId") Long postId,
+    @PostMapping("/{username}/{postId}")
+    public CommentDtoResponse addComment(@PathVariable String username,
+                                         @PathVariable Long postId,
                                          @Valid @RequestBody CommentDto commentDto) {
-        return commentService.addComment(commentDto, postId, username);
+        return commentServiceImpl.addComment(commentDto, postId, username);
     }
 
-    @DeleteMapping(path = "/{commentId}")
-    public ResponseEntity<?> deleteComment(@PathVariable("commentId") Long commentId) {
-        commentService.deleteComment(commentId);
+    @DeleteMapping("/{id}")
+    public ResponseEntity<?> deleteComment(@PathVariable Long id) {
+        commentServiceImpl.deleteComment(id);
         return ResponseEntity.ok().build();
     }
 }
