+++ src/main/java/com/example/eventinformationsystembackend/controller/EventController.java
@@ -2,100 +2,107 @@ package com.example.eventinformationsystembackend.controller;
 
 import com.example.eventinformationsystembackend.common.enums.EventType;
 import com.example.eventinformationsystembackend.dto.*;
-import com.example.eventinformationsystembackend.service.EventService;
+import com.example.eventinformationsystembackend.exception.RequiredPictureMissingException;
+import com.example.eventinformationsystembackend.service.implementation.EventServiceImpl;
 import jakarta.validation.Valid;
-import jakarta.websocket.server.PathParam;
-import org.apache.coyote.Response;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
-import org.springframework.security.access.annotation.Secured;
 import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.multipart.MultipartFile;
+import static com.example.eventinformationsystembackend.common.ExceptionMessages.*;
 
 import java.io.IOException;
 import java.util.List;
 
 @RestController
-@RequestMapping(path = "/events")
+@RequestMapping("/events")
 public class EventController {
 
-    private final EventService eventService;
+    private final EventServiceImpl eventServiceImpl;
 
     @Autowired
-    public EventController(EventService eventService) {
-        this.eventService = eventService;
+    public EventController(EventServiceImpl eventServiceImpl) {
+        this.eventServiceImpl = eventServiceImpl;
     }
 
-    @GetMapping(path = "/event-picture/{name}")
-    public ResponseEntity<?> getEventPicture(@PathVariable("name") String name) throws IOException {
-            byte[] eventPicture = eventService.getEventPicture(name);
+    @GetMapping("/event-picture/{id}")
+    public ResponseEntity<?> getEventPicture(@PathVariable Long id) throws IOException {
+            byte[] eventPicture = eventServiceImpl.getEventPicture(id);
             return ResponseEntity.status(HttpStatus.OK)
                     .contentType(MediaType.valueOf("image/png"))
                     .body(eventPicture);
     }
 
-    @GetMapping(path = "/{id}")
-    public EventDtoResponse getEventById(@PathVariable("id") Long id) {
-        return eventService.getEvent(id);
+    @GetMapping("/{id}")
+    public EventDtoResponse getEventById(@PathVariable Long id) {
+        return eventServiceImpl.getEvent(id);
     }
 
     @GetMapping
     public List<EventDtoResponse> getAllEvents(
             @RequestParam(value = "type", required = false) EventType type) {
-        return eventService.getAllEvents(type);
+        return eventServiceImpl.getAllEvents(type);
     }
 
     @GetMapping("/active")
     public int getNumberOfActiveEvents() {
-        return eventService.getNumberOfActiveEvents();
+        return eventServiceImpl.getNumberOfActiveEvents();
     }
 
     @GetMapping("/upcoming")
     public int getNumberOfUpcomingEvents(@RequestParam("type") int type) {
-        return eventService.getNumberOfUpcomingEvents(type);
+        return eventServiceImpl.getNumberOfUpcomingEvents(type);
     }
 
     @GetMapping("/booked")
     public int getNumberOfBookedEventsInThePast(@RequestParam("type") int type) {
-        return eventService.getNumberOfBookedEvents(type);
+        return eventServiceImpl.getNumberOfBookedEvents(type);
     }
 
     @GetMapping("/inactive")
     public int getNumberOfInactiveEvents() {
-        return eventService.getNumberOfInactiveEvents();
+        return eventServiceImpl.getNumberOfInactiveEvents();
     }
 
-    @GetMapping("/{eventId}/attendance")
-    public int getAttendancePercentageForEvent(@PathVariable("eventId") Long eventId) {
-        return eventService.getAttendancePercentageForEvent(eventId);
+    @GetMapping("/{id}/attendance")
+    public int getAttendancePercentageForEvent(@PathVariable Long id) {
+        return eventServiceImpl.getAttendancePercentageForEvent(id);
     }
 
     @PostMapping
     public EventDtoResponse addEvent(@RequestPart @Valid EventDto eventDto,
                                      @RequestPart(value = "eventPicture") MultipartFile eventPicture) {
-        return eventService.addEvent(eventDto, eventPicture);
+        if (eventPicture == null) {
+            throw new RequiredPictureMissingException(REQUIRED_PICTURE_IS_MISSING);
+        }
+
+        return eventServiceImpl.addEvent(eventDto, eventPicture);
     }
     
-    @PutMapping(path = "/{eventId}")
-    public EventDtoResponse updateEvent(@PathVariable("eventId") Long eventId,
+    @PutMapping("/{id}")
+    public EventDtoResponse updateEvent(@PathVariable Long id,
                                         @RequestPart @Valid EventDto eventDto,
                                         @RequestPart(value = "eventPicture") MultipartFile eventPicture) {
-        return eventService.updateEvent(eventId, eventDto, eventPicture);
+        if (eventPicture == null) {
+            throw new RequiredPictureMissingException(REQUIRED_PICTURE_IS_MISSING);
+        }
+
+        return eventServiceImpl.updateEvent(id, eventDto, eventPicture);
     }
 
     @PatchMapping("/activity-status/{id}")
-    public ResponseEntity<Void> toggleEventActivityStatus(@PathVariable("id") Long id) {
-        eventService.toggleEventActivityStatus(id);
+    public ResponseEntity<Void> toggleEventActivityStatus(@PathVariable Long id) {
+        eventServiceImpl.toggleEventActivityStatus(id);
         return ResponseEntity.ok().build();
     }
 
     @PreAuthorize("hasRole('ADMIN')")
-    @DeleteMapping(path = "/{eventId}")
-    public ResponseEntity<Void> deleteEvent(@PathVariable("eventId") Long eventId) {
-        eventService.deleteEvent(eventId);
+    @DeleteMapping(path = "/{id}")
+    public ResponseEntity<Void> deleteEvent(@PathVariable Long id) {
+        eventServiceImpl.deleteEvent(id);
         return ResponseEntity.ok().build();
     }
 }
