+++ src/main/java/com/example/eventinformationsystembackend/controller/OrderController.java
@@ -1,56 +1,52 @@
 package com.example.eventinformationsystembackend.controller;
 
-import com.example.eventinformationsystembackend.dto.OrderDto;
 import com.example.eventinformationsystembackend.dto.OrderDtoResponse;
 import com.example.eventinformationsystembackend.dto.OrderItemDtoResponse;
-import com.example.eventinformationsystembackend.model.OrderItem;
-import com.example.eventinformationsystembackend.service.OrderItemService;
-import com.example.eventinformationsystembackend.service.OrderService;
-import jakarta.validation.Valid;
+import com.example.eventinformationsystembackend.service.implementation.OrderItemServiceImpl;
+import com.example.eventinformationsystembackend.service.implementation.OrderServiceImpl;
 import jakarta.websocket.server.PathParam;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
 
 @RestController
-@RequestMapping(path = "/orders")
+@RequestMapping("/orders")
 public class OrderController {
-    private final OrderService orderService;
-    private final OrderItemService orderItemService;
+    private final OrderServiceImpl orderServiceImpl;
+    private final OrderItemServiceImpl orderItemServiceImpl;
 
     @Autowired
-    public OrderController(OrderService orderService,
-                           OrderItemService orderItemService) {
-        this.orderService = orderService;
-        this.orderItemService = orderItemService;
+    public OrderController(OrderServiceImpl orderServiceImpl,
+                           OrderItemServiceImpl orderItemServiceImpl) {
+        this.orderServiceImpl = orderServiceImpl;
+        this.orderItemServiceImpl = orderItemServiceImpl;
     }
 
-    @GetMapping(path = "/{id}")
-    public OrderDtoResponse getOrderById(@PathVariable("id") Long id) {
-        return orderService.getOrderById(id);
+    @GetMapping("/{id}")
+    public OrderDtoResponse getOrderById(@PathVariable Long id) {
+        return orderServiceImpl.getOrderById(id);
     }
 
-    @GetMapping(path = "/all/{username}")
-    public List<OrderDtoResponse> getAllOrdersForUser(@PathVariable("username") String username) {
-        return orderService.getAllOrdersForUsers(username);
+    @GetMapping("/all/{username}")
+    public List<OrderDtoResponse> getAllOrdersForUser(@PathVariable String username) {
+        return orderServiceImpl.getAllOrdersForUsers(username);
     }
 
-    @PostMapping(path = "/{username}")
-    public void createOrder(@PathVariable("username") String username,
+    @PostMapping("/{username}")
+    public void createOrder(@PathVariable String username,
                             @PathParam("couponCode") String couponCode) {
-        orderService.createOrder(username, couponCode);
+        orderServiceImpl.createOrder(username, couponCode);
     }
 
     //test
     @GetMapping(path = "/order-items/{orderId}")
-    public List<OrderItemDtoResponse> getOrderItems(@PathVariable("orderId") Long id) {
-        return orderItemService.getOrderItems(id);
+    public List<OrderItemDtoResponse> getOrderItems(@PathVariable Long id) {
+        return orderItemServiceImpl.getOrderItems(id);
     }
 
     @GetMapping("/prices/last/{ordersCount}")
-    public List<Integer> getPricesOfLastThirtyOrders(@PathVariable("ordersCount") int ordersCount) {
-        return orderService.getPricesOfLastSelectedOrders(ordersCount);
+    public List<Integer> getPricesOfLastThirtyOrders(@PathVariable int ordersCount) {
+        return orderServiceImpl.getPricesOfLastSelectedOrders(ordersCount);
     }
 }
