+++ src/main/java/com/example/eventinformationsystembackend/controller/SupportTicketController.java
@@ -2,9 +2,7 @@ package com.example.eventinformationsystembackend.controller;
 
 import com.example.eventinformationsystembackend.dto.SupportTicketDto;
 import com.example.eventinformationsystembackend.dto.SupportTicketDtoResponse;
-import com.example.eventinformationsystembackend.dto.SupportTicketReplyDto;
-import com.example.eventinformationsystembackend.dto.SupportTicketReplyDtoResponse;
-import com.example.eventinformationsystembackend.service.SupportTicketService;
+import com.example.eventinformationsystembackend.service.implementation.SupportTicketServiceImpl;
 import jakarta.validation.Valid;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.access.prepost.PreAuthorize;
@@ -13,38 +11,38 @@ import org.springframework.web.bind.annotation.*;
 import java.util.List;
 
 @RestController
-@RequestMapping(path = "/support-tickets")
+@RequestMapping("/support-tickets")
 public class SupportTicketController {
 
-    private final SupportTicketService supportTicketService;
+    private final SupportTicketServiceImpl supportTicketServiceImpl;
 
-    public SupportTicketController(SupportTicketService supportTicketService) {
-        this.supportTicketService = supportTicketService;
+    public SupportTicketController(SupportTicketServiceImpl supportTicketServiceImpl) {
+        this.supportTicketServiceImpl = supportTicketServiceImpl;
     }
 
     @GetMapping
     public List<SupportTicketDtoResponse> getAllSupportTickets() {
-        return supportTicketService.getAllSupportTickets();
+        return supportTicketServiceImpl.getAllSupportTickets();
     }
 
     @GetMapping("/{username}")
     public List<SupportTicketDtoResponse> getAllSupportTicketsForUser(
-            @PathVariable("username") String username) {
-        return supportTicketService.getAllSupportTicketsForUser(username);
+            @PathVariable String username) {
+        return supportTicketServiceImpl.getAllSupportTicketsForUser(username);
     }
 
     @PostMapping("/{username}")
     public SupportTicketDtoResponse createSupportTicket(
             @RequestBody @Valid SupportTicketDto supportTicketDto,
-            @PathVariable("username") String username) {
-        return supportTicketService.createSupportTicket(supportTicketDto,
+            @PathVariable String username) {
+        return supportTicketServiceImpl.createSupportTicket(supportTicketDto,
                 username);
     }
 
     @PreAuthorize("hasRole('ADMIN')")
     @DeleteMapping("/{id}")
-    public ResponseEntity<Void> deleteSupportTicket(@PathVariable("id") Long id) {
-        supportTicketService.deleteSupportTicket(id);
+    public ResponseEntity<Void> deleteSupportTicket(@PathVariable Long id) {
+        supportTicketServiceImpl.deleteSupportTicket(id);
         return ResponseEntity.ok().build();
     }
 }
