+++ src/main/java/com/example/eventinformationsystembackend/controller/SupportTicketReplyController.java
@@ -2,7 +2,7 @@ package com.example.eventinformationsystembackend.controller;
 
 import com.example.eventinformationsystembackend.dto.SupportTicketReplyDto;
 import com.example.eventinformationsystembackend.dto.SupportTicketReplyDtoResponse;
-import com.example.eventinformationsystembackend.service.SupportTicketReplyService;
+import com.example.eventinformationsystembackend.service.implementation.SupportTicketReplyServiceImpl;
 import jakarta.validation.Valid;
 import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.web.bind.annotation.*;
@@ -10,19 +10,19 @@ import org.springframework.web.bind.annotation.*;
 @RestController
 @RequestMapping("/support-ticket-replies")
 public class SupportTicketReplyController {
-    private SupportTicketReplyService supportTicketReplyService;
+    private final SupportTicketReplyServiceImpl supportTicketReplyServiceImpl;
 
-    public SupportTicketReplyController(SupportTicketReplyService supportTicketReplyService) {
-        this.supportTicketReplyService = supportTicketReplyService;
+    public SupportTicketReplyController(SupportTicketReplyServiceImpl supportTicketReplyServiceImpl) {
+        this.supportTicketReplyServiceImpl = supportTicketReplyServiceImpl;
     }
 
     @PreAuthorize("hasRole('ADMIN')")
     @PostMapping("/{username}")
     public SupportTicketReplyDtoResponse replyToSupportTicket(
             @Valid @RequestBody SupportTicketReplyDto supportTicketReplyDto,
-            @PathVariable("username") String username
+            @PathVariable String username
     ) {
-        return supportTicketReplyService.replyToSupportTicket(supportTicketReplyDto,
+        return supportTicketReplyServiceImpl.replyToSupportTicket(supportTicketReplyDto,
                 username);
     }
 }
