+++ src/main/java/com/example/eventinformationsystembackend/controller/UserController.java
@@ -4,8 +4,8 @@ import com.example.eventinformationsystembackend.dto.PasswordDto;
 import com.example.eventinformationsystembackend.dto.UserDto;
 import com.example.eventinformationsystembackend.dto.UserDtoResponse;
 import com.example.eventinformationsystembackend.dto.UserUpdateDto;
-import com.example.eventinformationsystembackend.service.ConfirmationTokenService;
-import com.example.eventinformationsystembackend.service.UserService;
+import com.example.eventinformationsystembackend.service.implementation.ConfirmationTokenServiceImpl;
+import com.example.eventinformationsystembackend.service.implementation.UserServiceImpl;
 import jakarta.validation.Valid;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
@@ -20,24 +20,24 @@ import java.io.IOException;
 import java.util.List;
 
 @RestController
-@RequestMapping(path = "/users")
+@RequestMapping("/users")
 public class UserController {
 
-    private final UserService userService;
-    private final ConfirmationTokenService confirmationTokenService;
+    private final UserServiceImpl userServiceImpl;
+    private final ConfirmationTokenServiceImpl confirmationTokenServiceImpl;
 
     @Autowired
-    public UserController(UserService userService,
-                          ConfirmationTokenService confirmationTokenService) {
-        this.userService = userService;
-        this.confirmationTokenService = confirmationTokenService;
+    public UserController(UserServiceImpl userServiceImpl,
+                          ConfirmationTokenServiceImpl confirmationTokenServiceImpl) {
+        this.userServiceImpl = userServiceImpl;
+        this.confirmationTokenServiceImpl = confirmationTokenServiceImpl;
     }
 
-    @GetMapping(path = "/confirmation")
+    @GetMapping("/confirmation")
     public RedirectView confirmToken(@RequestParam String token) {
         String confirmationStatus = "success";
 
-        if (!confirmationTokenService.confirmToken(token)) {
+        if (!confirmationTokenServiceImpl.confirmToken(token)) {
             confirmationStatus = "failure";
         }
 
@@ -49,18 +49,18 @@ public class UserController {
 
     @GetMapping
     public List<UserDtoResponse> getAllUsers() {
-        return userService.getAllUsers();
+        return userServiceImpl.getAllUsers();
     }
 
-    @GetMapping(path = "/{username}")
-    public UserDtoResponse getUser(@PathVariable("username") String username) {
-        return userService.getUser(username);
+    @GetMapping("/{username}")
+    public UserDtoResponse getUser(@PathVariable String username) {
+        return userServiceImpl.getUser(username);
     }
 
-    @GetMapping(path = "/profile-picture/{username}")
+    @GetMapping("/profile-picture/{username}")
     public ResponseEntity<?> getUserProfilePicture(
-            @PathVariable("username") String username) throws IOException {
-        byte[] profilePicture = userService.getUserProfilePicture(username);
+            @PathVariable String username) throws IOException {
+        byte[] profilePicture = userServiceImpl.getUserProfilePicture(username);
         return ResponseEntity.status(HttpStatus.OK)
                 .contentType(MediaType.valueOf("image/png"))
                 .body(profilePicture);
@@ -68,47 +68,47 @@ public class UserController {
 
     @PostMapping
     public UserDtoResponse registerUser(@RequestBody @Valid UserDto userDto) {
-        return userService.registerUser(userDto);
+        return userServiceImpl.registerUser(userDto);
     }
 
-    @PatchMapping(path = "/password/{username}")
-    public ResponseEntity<?> changePassword(@PathVariable("username") String username,
+    @PatchMapping("/password/{username}")
+    public ResponseEntity<?> changePassword(@PathVariable String username,
                                             @Valid @RequestBody PasswordDto passwordDto) {
-        userService.changePassword(username, passwordDto);
+        userServiceImpl.changePassword(username, passwordDto);
         return ResponseEntity.ok().build();
     }
 
-    @PatchMapping(path = "/profile-picture/default/{username}")
-    public ResponseEntity<?> resetProfilePictureToDefault(@PathVariable("username") String username) {
-        userService.resetProfilePictureToDefault(username);
+    @PatchMapping("/profile-picture/default/{username}")
+    public ResponseEntity<?> resetProfilePictureToDefault(@PathVariable String username) {
+        userServiceImpl.resetProfilePictureToDefault(username);
         return ResponseEntity.ok().build();
     }
 
-    @PutMapping(path = "/{username}")
-    public UserDtoResponse updateUser(@PathVariable("username") String username,
+    @PutMapping("/{username}")
+    public UserDtoResponse updateUser(@PathVariable String username,
                                       @RequestPart @Valid UserUpdateDto userUpdateDto,
                                       @RequestPart(required = false) MultipartFile profilePicture) {
-        return userService.updateUser(userUpdateDto, username, profilePicture);
+        return userServiceImpl.updateUser(userUpdateDto, username, profilePicture);
     }
 
     @PreAuthorize("hasRole('ADMIN')")
     @PatchMapping("/{username}/enabled")
-    public ResponseEntity<Void> toggleUserEnabledStatus(@PathVariable("username") String username) {
-        userService.toggleUserEnabledStatus(username);
+    public ResponseEntity<Void> toggleUserEnabledStatus(@PathVariable String username) {
+        userServiceImpl.toggleUserEnabledStatus(username);
         return ResponseEntity.ok().build();
     }
 
     @PreAuthorize("hasRole('ADMIN')")
     @PatchMapping("/{username}/locked")
     public ResponseEntity<Void> toggleUserLockedStatus(@PathVariable("username") String username) {
-        userService.toggleUserLockedStatus(username);
+        userServiceImpl.toggleUserLockedStatus(username);
         return ResponseEntity.ok().build();
     }
 
     @PreAuthorize("hasRole('ADMIN')")
     @DeleteMapping(path = "/{username}")
     public ResponseEntity<Void> deleteUser(@PathVariable("username") String username) {
-        userService.deleteUser(username);
+        userServiceImpl.deleteUser(username);
         return ResponseEntity.noContent().build();
     }
 }
