+++ src/main/java/com/example/eventinformationsystembackend/exception/GlobalExceptionHandler.java
@@ -64,6 +64,7 @@ public class GlobalExceptionHandler {
     }
 
     @ExceptionHandler({
+            InvalidCreditCardCredentialsException.class,
             CartItemTicketsExceedEventCapacity.class,
             PostDoesNotContainImageException.class,
             EmailAlreadyConfirmedException.class,
