+++ src/main/java/com/example/eventinformationsystembackend/exception/InvalidCreditCardCredentialsException.java
@@ -1,0 +1,7 @@
+package com.example.eventinformationsystembackend.exception;
+
+public class InvalidCreditCardCredentialsException extends RuntimeException {
+    public InvalidCreditCardCredentialsException(String message) {
+        super(message);
+    }
+}
