+++ src/main/java/com/example/eventinformationsystembackend/exception/RequiredPictureMissingException.java
@@ -1,0 +1,7 @@
+package com.example.eventinformationsystembackend.exception;
+
+public class RequiredPictureMissingException extends RuntimeException {
+    public RequiredPictureMissingException(String message) {
+        super(message);
+    }
+}
