+++ src/main/java/com/example/eventinformationsystembackend/security/SecurityConfiguration.java
@@ -34,10 +34,12 @@ public class SecurityConfiguration {
                 .cors(cors -> cors.configurationSource(corsConfigurationSource()))
                 .csrf(csrf -> csrf.disable())
                 .authorizeHttpRequests(auth -> auth
-                        .requestMatchers("/auth/**", "/posts/picture/*", "/tickets/verification/*").permitAll()
+                        .requestMatchers("/auth/**").permitAll()
+                        .requestMatchers("/tickets/verification/*").permitAll()
+                        .requestMatchers("/posts/{id:\\d+}/picture").permitAll()
                         .requestMatchers("/users", "/users/confirmation", "/users/profile-picture/*", "/users/register").permitAll()
                         .requestMatchers("/events", "/events/{id:\\d+}", "events/event-picture/{id:.+}", "events/concerts").permitAll()
-                        .requestMatchers("/artists", "/artists/profile-picture/{id:.+}").permitAll()
+                        .requestMatchers("/artists", "/artists/profile-picture/{id:\\d+}").permitAll()
                         .anyRequest().authenticated())
                 .sessionManagement(sess -> sess.sessionCreationPolicy(SessionCreationPolicy.STATELESS))
                 .authenticationProvider(authenticationProvider)
